Refactored TextEditor comp and also created EditingButtons comp for refactoring
