+++ src/components/card/modal/components/textEditor/TextEditor.tsx
@@ -1,11 +1,12 @@
 import { useEffect, useRef, useState } from "react";
 import DOMPurify from "dompurify";
-import { IoMdRedo, IoMdUndo } from "react-icons/io";
-import { VscListOrdered, VscListUnordered } from "react-icons/vsc";
-import "./TextEditor.scss";
+
+import { updateCard } from "../../../../../utils/api/updates";
 import { ICard } from "../../../../../types/board.type";
 import { IError } from "../../../../../types/status.type";
-import { updateCard } from "../../../../../utils/api/updates";
+
+import EditingButtons from "./EditingButtons";
+import "./TextEditor.scss";
 
 export default function TextEditor({ cardDetail }: { cardDetail: ICard }) {
   const [html, setHtml] = useState(DOMPurify.sanitize(cardDetail.description!));
@@ -21,6 +22,8 @@ export default function TextEditor({ cardDetail }: { cardDetail: ICard }) {
       elemRef.focus();
     }
   }, [isEditorialOpen]);
+
+  // end of hooks
 
   const handleSave = async () => {
     const token = localStorage.getItem("token");
@@ -46,10 +49,35 @@ export default function TextEditor({ cardDetail }: { cardDetail: ICard }) {
     setIsEditorialOpen(false);
   };
 
-  const handleTextEdit = (value: string) => {
-    document.execCommand(value);
-  };
+  const editedContent = (
+    <div className="text-editor__edited">
+      <div
+        className="text-editor__edited-content"
+        dangerouslySetInnerHTML={{ __html: `${localDescription}` }}
+      />
+      <button
+        onClick={() => {
+          setHtml(localDescription!);
+          setIsEditorialOpen(true);
+        }}
+      >
+        Edit
+      </button>
+    </div>
+  );
 
+  const editingCanvas = (
+    <p
+      className="text-editor__para-editing"
+      contentEditable
+      dangerouslySetInnerHTML={{
+        __html: html || "Write description here...",
+      }}
+      ref={paraRef}
+    ></p>
+  );
+
+  // JSX
   return (
     <div className="text-editor">
       {!localDescription && !isEditorialOpen ? (
@@ -62,52 +90,11 @@ export default function TextEditor({ cardDetail }: { cardDetail: ICard }) {
       ) : (
         <>
           {!isEditorialOpen ? (
-            <div className="text-editor__edited">
-              <div
-                className="text-editor__edited-content"
-                dangerouslySetInnerHTML={{ __html: `${localDescription}` }}
-              />
-              <button
-                onClick={() => {
-                  setHtml(localDescription!);
-                  setIsEditorialOpen(true);
-                }}
-              >
-                Edit
-              </button>
-            </div>
+            editedContent
           ) : (
             <div className="text-editor__editorial-area">
-              <div className="text-editor__btns">
-                <button onClick={() => handleTextEdit("bold")}>B</button>
-                <button onClick={() => handleTextEdit("italic")}>
-                  <em>I</em>
-                </button>
-                <button onClick={() => handleTextEdit("underline")}>
-                  <u>U</u>
-                </button>
-                <button onClick={() => handleTextEdit("insertOrderedList")}>
-                  <VscListOrdered />
-                </button>
-                <button onClick={() => handleTextEdit("insertUnorderedList")}>
-                  <VscListUnordered />
-                </button>
-                <button onClick={() => handleTextEdit("undo")}>
-                  <IoMdUndo size={20} />
-                </button>
-                <button onClick={() => handleTextEdit("redo")}>
-                  <IoMdRedo size={20} />
-                </button>
-              </div>
-
-              <p
-                className="text-editor__para-editing"
-                contentEditable
-                dangerouslySetInnerHTML={{
-                  __html: html || "Write description here...",
-                }}
-                ref={paraRef}
-              ></p>
+              <EditingButtons />
+              {editingCanvas}
               <div className="text-editor__save-btn">
                 <button onClick={handleSave}>Save</button>
                 <button onClick={handleCancel}>Cancel</button>
